make mesh rendering functional
