+++ src/canvas.h
@@ -21,12 +21,14 @@
 #include "part.h"
 #include "mesh.h"
 
+/*
 extern "C" {
     #define ANSI_DECLARATORS
     typedef double REAL;
     typedef void VOID; // required on linux (maybe not Windows?)
     #include "triangle.h"
 }
+*/
 
 class Canvas : public QOpenGLWidget, protected QOpenGLFunctions {
 public:
@@ -41,29 +43,13 @@ public:
     float theta = 45.0f; // camera angle, degrees
     float phi = M_PI/2.0f; // camera angle, degrees
     float zoom = 0.0005f;
+    //float zoom = 1.0f;
 
     glm::vec3 background_color = glm::vec3(0.1f, 0.1f, 0.1f);
     std::vector<std::shared_ptr<Part>>parts;
 
     QString filepath = "";
     QFileSystemWatcher* watcher;
-
-    QOpenGLVertexArrayObject VAO;
-    QOpenGLBuffer VBO;
-    QOpenGLShaderProgram shader;
-
-const char* vertex_source = "                           \n\
-    #version 330 core                                   \n\
-    layout (location = 0) in vec3 pos;                  \n\
-    void main(){                                        \n\
-        gl_Position = vec4(pos.xyz, 1.0f);  \n\
-    }";
-const char* fragment_source = "                         \n\
-    #version 330 core                                   \n\
-    out vec4 FragColor;                                 \n\
-    void main(){                                        \n\
-        FragColor = vec4(0.0f, 1.0f, 1.0f, 1.0f);            \n\
-    }";
 
     Canvas();
     ~Canvas();
